+++ components/.ordering
@@ -50,10 +50,6 @@ hook-http-client
 hook-http-server
 hook
 recorder-api
-recorder-process
-recorder-mocha
-recorder-jest
-recorder-remote
 recorder-cli
 receptor-file
 receptor-http
