+++ components/recorder-cli/default/.ordering
@@ -1,0 +1,5 @@
+process.mjs
+remote.mjs
+mocha.mjs
+jest.mjs
+index.mjs
