+++ components/recorder-cli/default/__fixture_jest__.mjs
@@ -1,3 +1,4 @@
+import { version } from "node:process";
 import "../../__fixture__.mjs";
 import * as Jest from "@jest/globals";
 import { convertFileUrlToPath } from "../../path/index.mjs";
@@ -5,34 +6,31 @@ import {
   createConfiguration,
   extendConfiguration,
 } from "../../configuration/index.mjs";
-import { main } from "./index.mjs";
+import { extendConfigurationNode } from "../../configuration-accessor/index.mjs";
+import { record } from "./jest.mjs";
 
 const { test: testJest } = Jest;
 
-const {
-  process: { version },
-} = globalThis;
-
 const configuration = createConfiguration("file:///w:/home/");
 
-for (const enabled of [true, false]) {
-  main(
+record(
+  extendConfigurationNode(
+    extendConfiguration(
+      configuration,
+      {
+        processes: [{ regexp: "", enabled: true }],
+        hooks: { cjs: false, esm: false, apply: false, http: false },
+        recorder: "jest",
+      },
+      "file:///w:/base/",
+    ),
     {
       cwd: () => convertFileUrlToPath("file:///w:/cwd"),
       argv: ["node", "main.mjs"],
       version,
     },
-    extendConfiguration(
-      configuration,
-      {
-        processes: [{ regexp: "", enabled }],
-        recorder: "process",
-        hooks: { cjs: false, esm: false, apply: false, http: false },
-      },
-      "file:///w:/base/",
-    ),
-  );
-}
+  ),
+);
 
 testJest("name1", () => {});
