+++ components/recorder-cli/default/process.mjs
@@ -1,0 +1,8 @@
+import { hook } from "../../hook/index.mjs";
+import { openAgent, recordStartTrack } from "../../agent/index.mjs";
+
+export const record = (configuration) => {
+  const agent = openAgent(configuration);
+  hook(agent, configuration);
+  recordStartTrack(agent, "process", {}, null);
+};
