--- components/recorder-jest/default/.env
@@ -1,2 +1,0 @@
-node
-test
