--- components/recorder-jest/default/.ordering
@@ -1,1 +1,0 @@
-index.mjs
