--- components/recorder-mocha/default/.env
@@ -1,2 +1,0 @@
-node
-test
